+++ src/app/pages/login/login.component.ts
@@ -60,9 +60,9 @@ export class LoginComponent implements OnInit, OnDestroy {
     {value: 'https://granska-api.sls.fi/', name: 'Staging'},
     {value: 'https://testa-api.sls.fi/', name: 'Development'},
     {value: 'https://testa-jansson-api.sls.fi/', name: 'Jansson testa'},
-    {value: 'https://testa-lukukirjat-api.sls.fi/', name: 'Lukukirja testa'},
-    {value: 'https://testa-parland-api.sls.fi/', name: 'Parland testa'},
-    {value: 'https://testa-westermarck-api.sls.fi/', name: 'Westermarck testa'},
+    {value: 'https://testa-lukukirjat-api.sls.fi/', name: 'Lukukirjat testa'},
+    {value: 'https://granska-parland-api.sls.fi/', name: 'Parland granska'},
+    {value: 'https://granska-westermarck-api.sls.fi/', name: 'Westermarck granska'},
     {value: ' ', name: 'Custom'}
   ];